+++ src/repositories/key_value.py
@@ -5,7 +5,7 @@ from sqlalchemy import select
 
 from src.model import KeyValue
 
-from .dao import CRUDRepository
+from .crud_repository import CRUDRepository
 
 
 class KeyValueRepository(CRUDRepository[KeyValue, str]):
