+++ src/repositories/major.py
@@ -4,7 +4,7 @@ from fastapi import Depends
 
 from src.model import Major
 
-from .dao import CRUDRepository
+from .crud_repository import CRUDRepository
 
 
 class MajorRepository(CRUDRepository[Major, int]):
